requirements.txt
asgiref==3.8.1
Django==5.0.3
django-phonenumber-field==7.3.0
phonenumbers==8.13.33
sqlparse==0.4.4
typing_extensions==4.10.0
